+++ src/components/Account/accountMenuSections.data.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { LayoutsContainer } from '@/containers/LayoutsContainer/LayoutsContainer';
-import { MyTariff } from '../../Tariff/MyTariff';
-import { AccountMenuItemsType } from '../accountTypes';
-import { MyProfile } from './MyProfile';
+import { MyTariff } from '../Tariff/MyTariff';
+import { AccountMenuItemsType } from './accountTypes';
+import { MyProfile } from '../MyProfile/MyProfile';
 
 export const ACCOUNT_MENU_ITEMS: AccountMenuItemsType[] = [
   {
