--- src/components/Forms/ProfileDataForm/profileDataFormDataTypes.ts
@@ -1,7 +1,0 @@
-export interface IProfileFormInput {
-  username: string;
-  birthday: string;
-  city: string;
-  password: string;
-  status: boolean;
-}
